:hammer: Moved components of GuidedTourStudentHomePageWidget.tsx into the /students/HomePage subdirectory
